% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{feature_importance}
\alias{feature_importance}
\title{Feature Importance}
\usage{
feature_importance(
  predictor,
  gexp,
  layer_gexp,
  normalize_gex = TRUE,
  n_cores = NULL
)
}
\arguments{
\item{predictor}{Predictor created by fit_predictor function}

\item{gexp}{Seurat object containing gene expression data}

\item{layer_gexp}{From which layer of the gexp Seurat Object should the gene expression data be extracted}

\item{normalize_gex}{Should gene expression levels be normalized across cells (see function gexp_normalize for details)}
}
\value{
Effect of the expression level of each gene on the prediction of the ADT values (averaged across all cells present in the input).
}
\description{
Only implemented for the case where z-score normalization was applied AFTER dimensionality reduction. \cr
Calculates the derivative of the predicted level of a given ADT as a function of the expression levels of a single gene 'dADT/dGene'\cr
Obtained by decomposing d(ADT)/d(Gene) = WJV where\cr
W = d(ADT)/d(zscores)\cr
J = d(zscores)/d(tSVD components)\cr
V = d(tSVD components)/d(gene expression)\cr
W and V directly correspond to the weights of the linear regression model & the right singular vectors of the tSVD.\cr
However, the effect of the z-score normalization had to be numerically approximated using the function 'jacobian' from the package 'numDeriv'.
}
