% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclinear.R
\name{filter_input_genes}
\alias{filter_input_genes}
\title{Forces a gene expression input matrix to match the gene expression input used in training of a given predictor.
Expression of genes which were not used in training the predictor is discarded, expression of genes used in training but not present in the input
are set to 0. Also reorders columns so they match the expected order for the tSVD projection}
\usage{
filter_input_genes(gexp, predictor)
}
\arguments{
\item{gexp}{gene expression matrix}

\item{predictor}{predictor object created by fit_predictor function}
}
\value{
gene expression matrix reshaped to match the set of genes used during training of the predictor (genes not present during training discarded, missing genes set to 0)
}
\description{
Forces a gene expression input matrix to match the gene expression input used in training of a given predictor.
Expression of genes which were not used in training the predictor is discarded, expression of genes used in training but not present in the input
are set to 0. Also reorders columns so they match the expected order for the tSVD projection
}
